Update BondDeviceProperties.java
Changing the addr variable to a "long" rather than "int" allows for devices with long addresses as specified by "kgailey
Kurt Gailey" at https://community.openhab.org/t/bond-home-bondhome-io/46654/47 .  Tested by modifying local file, building, and deploying jar to my environment; once done my devices went online and worked from my OpenHap 2.5.10-1 environment.

+++ bundles/org.openhab.binding.bondhome/src/main/java/org/openhab/binding/bondhome/internal/api/BondDeviceProperties.java
@@ -45,7 +45,7 @@ public class BondDeviceProperties {
     public boolean trustState;
     // The device address
     @Expose(serialize = true, deserialize = true)
-    public int addr;
+    public long addr;
     // The fan radio frequency
     @Expose(serialize = true, deserialize = true)
     public int freq;
